update to typescript
