+++ examples/04_react_router/src/index.tsx
@@ -1,0 +1,9 @@
+import * as React from "react";
+import { createRoot } from 'react-dom/client';
+
+import App from './App';
+
+const ele = document.getElementById('app');
+if (ele) {
+  createRoot(ele).render(<App />);
+}
