+++ examples/04_react_router/src/routerHistory.ts
@@ -1,4 +1,4 @@
-import { createBrowserHistory } from "history"; 
+import { createBrowserHistory } from 'history';
 
 const routerHistory = createBrowserHistory();
 export default routerHistory;
